+++ backend/src/main/java/com/greenfieldlibrary/backend/Persistence/BooksRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BooksRepository extends JpaRepository<Books, Long>{}    
+public interface BooksRepository extends JpaRepository<Books, Integer>{}
