+++ backend/src/main/java/com/greenfieldlibrary/backend/controllers/BooksResponse.java
@@ -1,13 +1,13 @@
 package com.greenfieldlibrary.backend.controllers;
 
 public class BooksResponse {
-    private Long idBooks;
+    private Integer idBooks;
     private String title;
     private String author;
     private String isbn;
     private String status;
 
-    public BooksResponse(Long idBooks, String title, String author, String isbn, String status) {
+    public BooksResponse(Integer idBooks, String title, String author, String isbn, String status) {
         this.idBooks = idBooks;
         this.title = title;
         this.author = author;
@@ -15,10 +15,10 @@ public class BooksResponse {
         this.status = status;
     }
     
-    public Long getIdBooks() {
+    public Integer getIdBooks() {
         return idBooks;
     }
-    public void setIdBooks(Long idBooks) {
+    public void setIdBooks(Integer idBooks) {
         this.idBooks = idBooks;
     }
     public String getTitle() {
